+++ src/PhpSpreadsheet/Calculation/Calculation.php
@@ -778,7 +778,7 @@ class Calculation
         ],
         'DAY' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\Day::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\DateParts::class, 'day'],
             'argumentCount' => '1',
         ],
         'DAYS' => [
@@ -1240,7 +1240,7 @@ class Calculation
         ],
         'HOUR' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\Hour::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\TimeParts::class, 'hour'],
             'argumentCount' => '1',
         ],
         'HYPERLINK' => [
@@ -1504,7 +1504,7 @@ class Calculation
         ],
         'ISOWEEKNUM' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\IsoWeekNum::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\Week::class, 'isoWeekNumber'],
             'argumentCount' => '1',
         ],
         'ISPMT' => [
@@ -1684,7 +1684,7 @@ class Calculation
         ],
         'MINUTE' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\Minute::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\TimeParts::class, 'minute'],
             'argumentCount' => '1',
         ],
         'MINVERSE' => [
@@ -1724,7 +1724,7 @@ class Calculation
         ],
         'MONTH' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\Month::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\DateParts::class, 'month'],
             'argumentCount' => '1',
         ],
         'MROUND' => [
@@ -1824,7 +1824,7 @@ class Calculation
         ],
         'NOW' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\Now::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\Current::class, 'now'],
             'argumentCount' => '0',
         ],
         'NPER' => [
@@ -2178,7 +2178,7 @@ class Calculation
         ],
         'SECOND' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\Second::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\TimeParts::class, 'second'],
             'argumentCount' => '1',
         ],
         'SEQUENCE' => [
@@ -2449,7 +2449,7 @@ class Calculation
         ],
         'TODAY' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\Today::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\Current::class, 'today'],
             'argumentCount' => '0',
         ],
         'TRANSPOSE' => [
@@ -2574,12 +2574,12 @@ class Calculation
         ],
         'WEEKDAY' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\WeekDay::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\Week::class, 'day'],
             'argumentCount' => '1,2',
         ],
         'WEEKNUM' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\WeekNum::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\Week::class, 'number'],
             'argumentCount' => '1,2',
         ],
         'WEIBULL' => [
@@ -2629,7 +2629,7 @@ class Calculation
         ],
         'YEAR' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,
-            'functionCall' => [DateTimeExcel\Year::class, 'evaluate'],
+            'functionCall' => [DateTimeExcel\DateParts::class, 'year'],
             'argumentCount' => '1',
         ],
         'YEARFRAC' => [
