+++ src/PhpSpreadsheet/Calculation/DateTime.php
@@ -63,15 +63,15 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\Now::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\Now class instead
+     * @See DateTimeExcel\Current::now()
+     *      Use the now method in the DateTimeExcel\Current class instead
      *
      * @return mixed Excel date/time serial value, PHP date/time serial value or PHP date/time object,
      *                        depending on the value of the ReturnDateType flag
      */
     public static function DATETIMENOW()
     {
-        return DateTimeExcel\Now::evaluate();
+        return DateTimeExcel\Current::now();
     }
 
     /**
@@ -90,15 +90,15 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\Today::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\Today class instead
+     * @See DateTimeExcel\Current::today()
+     *      Use the today method in the DateTimeExcel\Current class instead
      *
      * @return mixed Excel date/time serial value, PHP date/time serial value or PHP date/time object,
      *                        depending on the value of the ReturnDateType flag
      */
     public static function DATENOW()
     {
-        return DateTimeExcel\Today::evaluate();
+        return DateTimeExcel\Current::today();
     }
 
     /**
@@ -458,8 +458,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\Day::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\Day class instead
+     * @See DateTimeExcel\DateParts::day()
+     *      Use the day method in the DateTimeExcel\DateParts class instead
      *
      * @param mixed $dateValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard date string
@@ -468,7 +468,7 @@ class DateTime
      */
     public static function DAYOFMONTH($dateValue = 1)
     {
-        return DateTimeExcel\Day::evaluate($dateValue);
+        return DateTimeExcel\DateParts::day($dateValue);
     }
 
     /**
@@ -482,8 +482,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\WeekDay::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\WeekDay class instead
+     * @See DateTimeExcel\Week::day()
+     *      Use the day method in the DateTimeExcel\Week class instead
      *
      * @param float|int|string $dateValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard date string
@@ -496,7 +496,7 @@ class DateTime
      */
     public static function WEEKDAY($dateValue = 1, $style = 1)
     {
-        return DateTimeExcel\WeekDay::evaluate($dateValue, $style);
+        return DateTimeExcel\Week::day($dateValue, $style);
     }
 
     /**
@@ -687,8 +687,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\WeekNum::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\WeekNum class instead
+     * @See DateTimeExcel\Week::number(()
+     *      Use the number method in the DateTimeExcel\Week class instead
      *
      * @param mixed $dateValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard date string
@@ -708,7 +708,7 @@ class DateTime
      */
     public static function WEEKNUM($dateValue = 1, $method = self::STARTWEEK_SUNDAY)
     {
-        return DateTimeExcel\WeekNum::evaluate($dateValue, $method);
+        return DateTimeExcel\Week::number($dateValue, $method);
     }
 
     /**
@@ -721,8 +721,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\IsoWeekNum::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\IsoWeekNum class instead
+     * @See DateTimeExcel\Week::isoWeekNumber()
+     *      Use the isoWeekNumber method in the DateTimeExcel\Week class instead
      *
      * @param mixed $dateValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard date string
@@ -731,7 +731,7 @@ class DateTime
      */
     public static function ISOWEEKNUM($dateValue = 1)
     {
-        return DateTimeExcel\IsoWeekNum::evaluate($dateValue);
+        return DateTimeExcel\Week::isoWeekNumber($dateValue);
     }
 
     /**
@@ -745,8 +745,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\Month::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\Month class instead
+     * @See DateTimeExcel\DateParts::month()
+     *      Use the month method in the DateTimeExcel\DateParts class instead
      *
      * @param mixed $dateValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard date string
@@ -755,7 +755,7 @@ class DateTime
      */
     public static function MONTHOFYEAR($dateValue = 1)
     {
-        return DateTimeExcel\Month::evaluate($dateValue);
+        return DateTimeExcel\DateParts::month($dateValue);
     }
 
     /**
@@ -769,8 +769,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\Year::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\Year class instead
+     * @See DateTimeExcel\DateParts::year()
+     *      Use the ear method in the DateTimeExcel\DateParts class instead
      *
      * @param mixed $dateValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard date string
@@ -779,7 +779,7 @@ class DateTime
      */
     public static function YEAR($dateValue = 1)
     {
-        return DateTimeExcel\Year::evaluate($dateValue);
+        return DateTimeExcel\DateParts::year($dateValue);
     }
 
     /**
@@ -793,8 +793,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\Hour::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\Hour class instead
+     * @See DateTimeExcel\TimeParts::hour()
+     *      Use the hour method in the DateTimeExcel\TimeParts class instead
      *
      * @param mixed $timeValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard time string
@@ -803,7 +803,7 @@ class DateTime
      */
     public static function HOUROFDAY($timeValue = 0)
     {
-        return DateTimeExcel\Hour::evaluate($timeValue);
+        return DateTimeExcel\TimeParts::hour($timeValue);
     }
 
     /**
@@ -817,8 +817,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\Minute::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\Minute class instead
+     * @See DateTimeExcel\TimeParts::minute()
+     *      Use the minute method in the DateTimeExcel\TimeParts class instead
      *
      * @param mixed $timeValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard time string
@@ -827,7 +827,7 @@ class DateTime
      */
     public static function MINUTE($timeValue = 0)
     {
-        return DateTimeExcel\Minute::evaluate($timeValue);
+        return DateTimeExcel\TimeParts::minute($timeValue);
     }
 
     /**
@@ -841,8 +841,8 @@ class DateTime
      *
      * @Deprecated 1.18.0
      *
-     * @See DateTimeExcel\Second::evaluate()
-     *      Use the evaluate method in the DateTimeExcel\Second class instead
+     * @See DateTimeExcel\TimeParts::second()
+     *      Use the second method in the DateTimeExcel\TimeParts class instead
      *
      * @param mixed $timeValue Excel date serial value (float), PHP date timestamp (integer),
      *                                    PHP DateTime object, or a standard time string
@@ -851,7 +851,7 @@ class DateTime
      */
     public static function SECOND($timeValue = 0)
     {
-        return DateTimeExcel\Second::evaluate($timeValue);
+        return DateTimeExcel\TimeParts::second($timeValue);
     }
 
     /**
