+++ tests/PhpSpreadsheetTests/Calculation/Functions/DateTime/WeekDayTest.php
@@ -2,7 +2,7 @@
 
 namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\DateTime;
 
-use PhpOffice\PhpSpreadsheet\Calculation\DateTimeExcel\WeekDay;
+use PhpOffice\PhpSpreadsheet\Calculation\DateTimeExcel\Week;
 
 class WeekDayTest extends AllSetupTeardown
 {
@@ -28,8 +28,8 @@ class WeekDayTest extends AllSetupTeardown
     public function testWEEKDAYwith1904Calendar(): void
     {
         self::setMac1904();
-        self::assertEquals(7, WeekDay::evaluate('1904-01-02'));
-        self::assertEquals(6, WeekDay::evaluate('1904-01-01'));
-        self::assertEquals(6, WeekDay::evaluate(null));
+        self::assertEquals(7, Week::day('1904-01-02'));
+        self::assertEquals(6, Week::day('1904-01-01'));
+        self::assertEquals(6, Week::day(null));
     }
 }
